// Portal SAE/frontend/src/app/interfaces/account/updateAccount.dto.ts
export interface UpdateAccountDTO {
  numeroIdentificacion?: string;
  email?: string;
  paginaWeb?: string;
  codigoPais?: string;
  telefonoFijo?: string;
  telefonoSecundario?: string;
  sectorEconomico?: string;
  pais?: string;
  departamento?: string;
  ciudad?: string;
  numeroEmpleados?: number;

  tipoIdentificacion?: string;
  tipoServicio?: string;
  codigoUsuario?: string;
  codigoConsultorAsignado?: string;
}