// Portal SAE/frontend/src/app/interfaces/account/account.dto.ts
export interface AccountDTO {
  codigo: number;
  numeroIdentificacion: string;
  email: string;
  paginaWeb: string;
  codigoPais: string;
  telefonoFijo: string;
  telefonoSecundario: string;
  sectorEconomico: string;
  pais: string;
  departamento: string;
  ciudad: string;
  numeroEmpleados: number;

  tipoIdentificacion: string;
  tipoServicio: string;
  codigoUsuario: string;
  codigoConsultorAsignado: string;
}